backend/requirements.txt
# AuraGraph backend dependencies
# Team Wowffulls | IIT Roorkee | AI Study Buddy Challenge

# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# ── Pydantic ───────────────────────────────────────────────────────────────────
pydantic>=2.0.0,<2.10.0
pydantic-settings>=2.0.0,<2.7.0

# ── Azure OpenAI / Semantic Kernel ─────────────────────────────────────────────
semantic-kernel>=1.3.0
openai>=1.0.0                        # AzureOpenAI client used directly in pipeline

# ── Azure AI Services (SDK) ───────────────────────────────────────────────────
azure-ai-vision-imageanalysis>=1.0.0  # Azure AI Vision: slide OCR + figure captions
azure-ai-contentsafety>=1.0.0         # Azure AI Content Safety: screens all LLM output
azure-search-documents>=11.4.0        # Azure AI Search: vector RAG for textbook chunks
azure-identity>=1.15.0                # DefaultAzureCredential (managed identity support)

# ── Azure Cosmos DB ────────────────────────────────────────────────────────────
azure-cosmos>=4.6.0                   # Azure Cosmos DB: persistent knowledge graph store

# ── Document extraction ────────────────────────────────────────────────────────
PyPDF2>=3.0.1
pdfplumber>=0.11.0
python-pptx>=1.0.0
pymupdf>=1.24.0

# ── HTTP ──────────────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── Image processing ──────────────────────────────────────────────────────────
Pillow>=10.0.0
pillow-heif>=0.13.0

# ── Numerics (TF-IDF fallback embedder) ───────────────────────────────────────
numpy>=1.24.0

# ── Auth ──────────────────────────────────────────────────────────────────────
bcrypt>=4.0.0

# ── Media ingestion ───────────────────────────────────────────────────────────
yt-dlp>=2024.1.0                     # YouTube/video transcript extraction
